// web/src/api/models/TypesAPIKey.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypesAPIKey
 */
export interface TypesAPIKey {
    /**
     * 
     * @type {number}
     * @memberof TypesAPIKey
     */
    created_at: number;
    /**
     * 
     * @type {string}
     * @memberof TypesAPIKey
     */
    created_by: string;
    /**
     * 
     * @type {string}
     * @memberof TypesAPIKey
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof TypesAPIKey
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof TypesAPIKey
     */
    prefix: string;
    /**
     * 
     * @type {number}
     * @memberof TypesAPIKey
     */
    revoked_at: number;
}

/**
 * Check if a given object implements the TypesAPIKey interface.
 */
export function instanceOfTypesAPIKey(value: object): value is TypesAPIKey {
    if (!('created_at' in value) || value['created_at'] === undefined) return false;
    if (!('created_by' in value) || value['created_by'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('prefix' in value) || value['prefix'] === undefined) return false;
    if (!('revoked_at' in value) || value['revoked_at'] === undefined) return false;
    return true;
}

export function TypesAPIKeyFromJSON(json: any): TypesAPIKey {
    return TypesAPIKeyFromJSONTyped(json, false);
}

export function TypesAPIKeyFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesAPIKey {
    if (json == null) {
        return json;
    }
    return {
        
        'created_at': json['created_at'],
        'created_by': json['created_by'],
        'id': json['id'],
        'name': json['name'],
        'prefix': json['prefix'],
        'revoked_at': json['revoked_at'],
    };
}

export function TypesAPIKeyToJSON(json: any): TypesAPIKey {
    return TypesAPIKeyToJSONTyped(json, false);
}

export function TypesAPIKeyToJSONTyped(value?: TypesAPIKey | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'created_at': value['created_at'],
        'created_by': value['created_by'],
        'id': value['id'],
        'name': value['name'],
        'prefix': value['prefix'],
        'revoked_at': value['revoked_at'],
    };
}